Move generic components to appropriate files
